% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosterR6.R
\name{Booster}
\alias{Booster}
\title{R6 Class for Gradient Boosting}
\description{
An R6 Class that provides an interface to gradient boosting with neural network feature transformation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{estimators}}{List of fitted base models}

\item{\code{learning_rate}}{Learning rate for boosting}

\item{\code{losses}}{Vector of training losses}

\item{\code{n_estimators}}{Number of estimators used}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Booster-new}{\code{Booster$new()}}
\item \href{#method-Booster-fit}{\code{Booster$fit()}}
\item \href{#method-Booster-predict}{\code{Booster$predict()}}
\item \href{#method-Booster-clone}{\code{Booster$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Booster-new"></a>}}
\if{latex}{\out{\hypertarget{method-Booster-new}{}}}
\subsection{Method \code{new()}}{
Create a new Booster object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Booster$new(
  model_name = "ExtraTreeRegressor",
  n_estimators = 100L,
  learning_rate = 0.1,
  tolerance = 1e-04,
  calibration = FALSE,
  seed = 123L,
  show_progress = TRUE,
  verbose = FALSE,
  venv_path = "./venv"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the base model}

\item{\code{n_estimators}}{Number of boosting iterations}

\item{\code{learning_rate}}{Learning rate for boosting}

\item{\code{tolerance}}{Convergence tolerance}

\item{\code{calibration}}{Whether to calibrate the model}

\item{\code{seed}}{Random seed}

\item{\code{show_progress}}{Whether to show progress bar}

\item{\code{verbose}}{Whether to print detailed output}

\item{\code{venv_path}}{Path to the virtual environment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Booster-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Booster-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the boosting model to training data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Booster$fit(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Feature matrix}

\item{\code{y}}{Target vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The fitted object (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Booster-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Booster-predict}{}}}
\subsection{Method \code{predict()}}{
Make predictions on new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Booster$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{New data to predict on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of predictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Booster-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Booster-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Booster$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
