% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list.R
\name{get_model_list}
\alias{get_model_list}
\title{Get a list of all models in scikit-learn}
\usage{
get_model_list(venv_path = "venv")
}
\value{
A list with two elements:
  - `classifiers`: A character vector of all classifier models
  - `regressors`: A character vector of all regressor models
}
\description{
This function retrieves a list of all models available in scikit-learn.
It imports the necessary Python modules and retrieves all estimators,
filtering them into classifiers and regressors.
}
\examples{

# model_list <- get_model_list()
# print(model_list$classifiers)

}
