% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booster.R
\name{predict.booster}
\alias{predict.booster}
\title{Predict using a boosted model}
\usage{
\method{predict}{booster}(object, newdata, ...)
}
\arguments{
\item{object}{A boosted model object}

\item{newdata}{New data to predict on}

\item{...}{Additional arguments}
}
\description{
Predict using a boosted model
}
