% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{predict.regressor}
\alias{predict.regressor}
\title{Predict method for regressor objects}
\usage{
\method{predict}{regressor}(
  object,
  newdata,
  nsim = 250L,
  level = 95,
  method = c("none", "splitconformal", "surrogate", "bootstrap", "tsbootstrap"),
  seed = 123,
  ...
)
}
\arguments{
\item{object}{A regressor object}

\item{newdata}{New data to predict on}

\item{nsim}{Number of simulations for bootstrap/tsbootstrap}

\item{level}{Confidence level for prediction intervals}

\item{method}{Method for computing prediction intervals}

\item{...}{Additional arguments}
}
\description{
Predict method for regressor objects
}
