% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearnR.R
\name{predict_model}
\alias{predict_model}
\title{Make predictions using a trained model}
\usage{
predict_model(X_new, model_id)
}
\arguments{
\item{X_new}{New feature matrix for predictions}

\item{model_id}{Model ID (path) returned by train_model}
}
\value{
Vector of predictions
}
\description{
Make predictions using a trained model
}
