% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictBoosterCpp}
\alias{predictBoosterCpp}
\title{Predict using a boosted model}
\usage{
predictBoosterCpp(booster, x)
}
\arguments{
\item{booster}{A boosted model object}

\item{x}{New data to predict on}
}
\description{
Predict using a boosted model
}
