% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressor.R
\name{regressor}
\alias{regressor}
\title{Create a regression model using tisthemachinelearner}
\usage{
regressor(x, y, model_name, ...)
}
\arguments{
\item{x}{Feature matrix}

\item{y}{Target vector}

\item{model_name}{Name of the sklearn model to use}

\item{...}{Additional parameters passed to the sklearn model}
}
\value{
A regressor object
}
\description{
Create a regression model using tisthemachinelearner
}
