% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_sklearn}
\alias{setup_sklearn}
\title{Setup Python environment using uv}
\usage{
setup_sklearn(venv_path = "venv")
}
\arguments{
\item{venv_path}{Path to virtual environment (default: "./venv")}
}
\description{
Setup Python environment using uv
}
\examples{
\dontrun{
# After creating venv with: uv venv venv
setup_sklearn()
}
}
