% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sklearnR.R
\name{train_model}
\alias{train_model}
\title{Train a machine learning model}
\usage{
train_model(X, y, model_name, params = list())
}
\arguments{
\item{X}{Feature matrix}

\item{y}{Target vector}

\item{model_name}{Name of the sklearn model to use (e.g., "RandomForestRegressor")}

\item{params}{List of parameters for the model}
}
\value{
A model ID string (path to saved model)
}
\description{
Train a machine learning model
}
