% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booster.R
\name{booster}
\alias{booster}
\title{Fit a boosting model with neural network feature transformation}
\usage{
booster(
  x,
  y,
  model_name,
  n_estimators = 100L,
  learning_rate = 0.1,
  tolerance = 1e-04,
  calibration = FALSE,
  seed = 123L,
  show_progress = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Input matrix}

\item{y}{Target vector}

\item{model_name}{Name of the base model}

\item{n_estimators}{Number of boosting iterations}

\item{learning_rate}{Learning rate for boosting}

\item{calibration}{Whether to calibrate the model}

\item{seed}{Random seed}

\item{...}{Additional arguments passed to the regressor}
}
\description{
Fit a boosting model with neural network feature transformation
}
