% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorR6.R, R/regressor_class.R
\name{Regressor}
\alias{Regressor}
\title{R6 Class for Scikit-learn Regressors}
\description{
An R6 Class that provides an interface to scikit-learn regression models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The underlying sklearn model}

\item{\code{residuals}}{Model residuals}

\item{\code{df.residual}}{Degrees of freedom of residuals}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{residuals}}{Model residuals}

\item{\code{df.residual}}{Degrees of freedom of residuals}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Regressor-new}{\code{Regressor$new()}}
\item \href{#method-Regressor-fit}{\code{Regressor$fit()}}
\item \href{#method-Regressor-predict}{\code{Regressor$predict()}}
\item \href{#method-Regressor-clone}{\code{Regressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$new(model_name = "LinearRegression", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$fit(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$predict(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The fitted model object}

\item{\code{model_name}}{Name of the model to use}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Regressor-new}{\code{Regressor$new()}}
\item \href{#method-Regressor-fit}{\code{Regressor$fit()}}
\item \href{#method-Regressor-predict}{\code{Regressor$predict()}}
\item \href{#method-Regressor-clone}{\code{Regressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-new}{}}}
\subsection{Method \code{new()}}{
Create a new Regressor instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$new(model_name = "LinearRegression", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use}

\item{\code{...}}{Additional parameters for the model}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Regressor` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model to training data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$fit(x, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Feature matrix}

\item{\code{y}}{Target variable}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self (for method chaining)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-predict}{}}}
\subsection{Method \code{predict()}}{
Make predictions on new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$predict(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Feature matrix for predictions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of predictions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
