% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressorR6.R
\name{Regressor}
\alias{Regressor}
\title{R6 Class for Scikit-learn Regressors}
\description{
An R6 Class that provides an interface to scikit-learn regression models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The underlying sklearn model}

\item{\code{residuals}}{Model residuals}

\item{\code{df.residual}}{Degrees of freedom of residuals}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Regressor-new}{\code{Regressor$new()}}
\item \href{#method-Regressor-fit}{\code{Regressor$fit()}}
\item \href{#method-Regressor-predict}{\code{Regressor$predict()}}
\item \href{#method-Regressor-clone}{\code{Regressor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-new}{}}}
\subsection{Method \code{new()}}{
Create a new Regressor object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$new(model_name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the sklearn model to use}

\item{\code{...}}{Additional parameters passed to the sklearn model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the model to training data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$fit(x, y, calibration = FALSE, seed = 42L, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Feature matrix}

\item{\code{y}}{Target vector}

\item{\code{calibration}}{Logical flag to indicate if calibration of residuals should be used}

\item{\code{seed}}{Seed for random number generator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-predict}{}}}
\subsection{Method \code{predict()}}{
Make predictions on new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$predict(
  newdata,
  method = c("none", "splitconformal", "surrogate", "bootstrap", "tsbootstrap",
    "bayesian"),
  nsim = 250L,
  level = 95,
  seed = 123
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{New data to predict on}

\item{\code{method}}{Method for computing prediction intervals}

\item{\code{nsim}}{Number of simulations for bootstrap/tsbootstrap}

\item{\code{level}}{Confidence level for prediction intervals}

\item{\code{seed}}{Random seed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Regressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
